% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_diff_means.R
\name{ci_diff_means}
\alias{ci_diff_means}
\title{Confidence interval for difference in means}
\usage{
ci_diff_means(data, var, by)
}
\arguments{
\item{data}{A data frame}

\item{var}{A numeric variable containing the data for the calculation of the means}

\item{by}{A grouping variable with only two values}
}
\value{
A numeric value for the difference in means
}
\description{
Calulates a 95% CI for the difference between two means.
Uses an approximation to degrees of freedom of n observations - 2.
Thus, may give broader confidence intervals than built in \code{t.test()} function
}
\examples{
data(sleep)
sleep$group <- as.numeric(sleep$group)
ci_diff_means(data = sleep, var = "extra", by = "group")
}
