% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_dataframes.R
\name{clear_dataframes}
\alias{clear_dataframes}
\title{removes data frames from global environment}
\usage{
clear_dataframes()
}
\description{
For those occasions when you want to clear all your data frames out of memory without also removing any functions you may have loaded in.
Adapted from \url{http://stackoverflow.com/questions/19684819/get-list-of-available-data-frames}
}
\examples{
dat <- as.data.frame(x = rnorm(3), y = rnorm(3))
ls()
clear_dataframes()
ls()
}
