% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_dataframes.R
\name{clear_vectors}
\alias{clear_vectors}
\title{removes vectors from global environment}
\usage{
clear_vectors()
}
\description{
For those times when you want to clear up your global environment and get rid
of all the vectors.
}
\examples{
vec <- c(1:3)
ls()
clear_vectors()
ls()
}
