% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_diff_means.R
\name{nice_diff_means}
\alias{nice_diff_means}
\title{Difference in means}
\usage{
nice_diff_means(data, var, by)
}
\arguments{
\item{data}{A data frame}

\item{var}{A numeric variable containing the data for the calculation of the means}

\item{by}{A grouping variable with only two values}
}
\value{
A numeric value for the difference in means
}
\description{
Calculates a difference in means between two groups
}
\examples{
data(sleep)
sleep$group <- as.numeric(sleep$group)
nice_diff_means(data = sleep, var = "extra", by = "group")
}
