% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_things.R
\name{nice_govuk_date}
\alias{nice_govuk_date}
\title{Formats dates to gov.uk requirements}
\usage{
nice_govuk_date(x)
}
\arguments{
\item{x}{A date object}
}
\value{
A string
}
\description{
gov.uk dates are supposed to be formatted DD Month YYYY
This function takes a date and reformats it accordingly
}
\examples{
nice_govuk_date(Sys.Date())
}
