% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_things.R
\name{nice_inline_r}
\alias{nice_inline_r}
\title{Allows user to write examples of inline R code.}
\usage{
nice_inline_r(x)
}
\arguments{
\item{x}{A string of R code for printing as code, inline}
}
\value{
A formatted string
}
\description{
Yihui gets the credit for this one.
\url{http://stackoverflow.com/questions/20409172/how-to-display-verbatim-inline-r-code-with-backticks-using-rmarkdown}
Writing text to give examples of inline r code is tricky.
This function should achieve that.
}
\examples{
nice_inline_r("round(3.77155454, 3)")
}
