% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_median.R
\name{nice_median}
\alias{nice_median}
\title{nice_median}
\usage{
nice_median(x, digits = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{The number of digits to which values will be rounded}
}
\value{
A median with interquartile range in brackets
}
\description{
Prints a median with interquartile range.
}
\examples{
data(mtcars)
nice_median(mtcars$cyl)
}
