% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_date_functions.R
\name{nice_month_string}
\alias{nice_month_string}
\title{nice_month_factor}
\usage{
nice_month_string(x, abbreviated_out = FALSE, year_type = "calendar")
}
\arguments{
\item{x}{A month string, will take either abbreviated or full length month strings}

\item{abbreviated_out}{Should the output be abbreviated? default = FALSE}

\item{year_type}{string, either calendar or financial (year starts April)}
}
\value{
A factor with months in order as the occur in the year.
}
\description{
One of the problems with the Gregorian calendar is that months are not alphabetical.
In R the default sorting of strings is in alphabetical order.
For tables and graphs of data by month this may mean repeated creation of month variables, which is tedious.
}
\examples{
nice_month_string(x = "Dec", abbreviated_out = FALSE, year_type = "calendar")
}
