% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_things.R
\name{nice_pc_change}
\alias{nice_pc_change}
\title{Easily calculate a percentage change}
\usage{
nice_pc_change(start, end)
}
\arguments{
\item{start}{The value at the start}

\item{end}{The value at the end}
}
\value{
A percentage change from start to end
}
\description{
Intended for inline printing, allows easy calculation of percentage changes
}
\examples{
nice_pc_change(5, 10)
}
