% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_things.R
\name{nice_perc}
\alias{nice_perc}
\title{Nicely formats percentages for printing}
\usage{
nice_perc(numerator, denominator)
}
\arguments{
\item{numerator}{The numerator}

\item{denominator}{The denominator}
}
\value{
A string giving the percentage to one decimal place
}
\description{
Takes a denominator and a numerator, calculates a percentage and formats for printing in a markdown doc.
}
\examples{
nice_perc(5, 10)
}
