% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_things.R
\name{nice_pval}
\alias{nice_pval}
\title{Formats p values for nice printing}
\usage{
nice_pval(x)
}
\arguments{
\item{x}{A p-value}
}
\value{
A string where the p-value is formatted to three decimal places or "<0.001" where appropriate
}
\description{
Takes a p value and formats it to three decimal places for printing.
}
\examples{
nice_pval(0.045)
nice_pval(1e-05)
}
