% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_risk_difference.R
\name{nice_risk_difference}
\alias{nice_risk_difference}
\title{nice_risk_difference}
\usage{
nice_risk_difference(n1, d1, n2, d2)
}
\arguments{
\item{n1}{Numerator 1}

\item{d1}{Denominator 1}

\item{n2}{Numerator 2}

\item{d2}{Denominator 2}
}
\description{
calculates a risk difference and confidence interval with a 95% confidence interval
}
\examples{
# table 16.2 Kirkwood and Sterne
nice_risk_difference(n1 = 20, d1 = 240, n2 = 80, d2 = 220)
}
