% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_rr.R
\name{nice_rr}
\alias{nice_rr}
\title{nice_rr}
\usage{
nice_rr(n1, d1, n2, d2)
}
\arguments{
\item{n1}{numerator for the reference group}

\item{d1}{denominator for the reference group}

\item{n2}{numerator for the comparator group}

\item{d2}{denominator for the comparator group}
}
\description{
Calculate risk ratios from line-list data
}
\examples{
data(smokers)
lc_agg <- smokers \%>\%
  dplyr::group_by(smoker) \%>\%
  dplyr::summarise(denominator = dplyr::n(), numerator = sum(lung_cancer)) \%>\%
  tidyr::pivot_wider(names_from = smoker, values_from = c(denominator, numerator))
lc_agg
nice_rr(n1 = lc_agg$numerator_0, d1 = lc_agg$denominator_0,
    n2 = lc_agg$numerator_1, d2 = lc_agg$denominator_1)
}
