% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_things.R
\name{nice_rversionstring}
\alias{nice_rversionstring}
\title{nice_rversionstring}
\usage{
nice_rversionstring()
}
\value{
A character string giving the R version
}
\description{
Print a nice R version string without the date at the end
}
