% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicethings-package.R
\docType{package}
\name{nicethings-package}
\alias{nicethings}
\alias{nicethings-package}
\title{nicethings: Nicely formats things for printing inline in Rmarkdown documents and other nice things}
\description{
Printing R output inline to markdown documents can be long winded and use many nested functions. This makes it hard to write the code and hard to read it later on. This package aims to simplify that by nicely wrapping up these nested functions to make writing inline code nicer. It may also do other nice things that I find useful.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonthelwall/nicethings}
  \item \url{https://simonthelwall.github.io/nicethings/}
  \item Report bugs at \url{https://github.com/simonthelwall/nicethings/issues}
}

}
\author{
\strong{Maintainer}: Simon Thelwall \email{simon.thelwall@phe.gov.uk} (\href{https://orcid.org/0000-0002-0434-2724}{ORCID})

}
\keyword{internal}
