% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_utilities.R
\name{pipe_nice_names}
\alias{pipe_nice_names}
\title{Format variable names into standard format in a pipe.}
\usage{
pipe_nice_names(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A data frame with nicely formatted names
}
\description{
Turns column names of data frames into standardised format.
Dots are replaced with underscores, trailing underscores are removed and everything is made lower case.
Punctuation and spaces are also replaced by underscores and double underscores are removed last
}
\examples{
\dontrun{
data(mtcars)
names(mtcars) <- toupper(names(mtcars))
mtcars \%>\% pipe_nice_names()
}
}
