% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_confint.R
\name{rate_confint}
\alias{rate_confint}
\title{Confidence interval for a rate calculation}
\usage{
rate_confint(numerator, denominator)
}
\arguments{
\item{numerator}{Count of events}

\item{denominator}{Count of person-years}
}
\value{
A data frame with three columns giving the rate, lower confidence interval and upper confidence interval.
}
\description{
Calculates a 95 percent confidence interval for a rate using the Rothman/Greenland method described in \url{http://www.openepi.com/PDFDocs/PersonTime1Doc.pdf}.
}
\examples{
dat <- data.frame(n = 1, denom = 10)
rate_confint(dat$n, dat$denom)
dat <- data.frame(n = 5, denom = 25) # replicates output from openepi just fine.
rate_confint(dat$n, dat$denom)
}
