% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_confint.R
\name{rate_difference}
\alias{rate_difference}
\title{Confidence interval for a rate difference}
\usage{
rate_difference(numerator1, denominator1, numerator2, denominator2)
}
\arguments{
\item{numerator1}{The numerator for the first level for comparison (e.g. the unexposed group)}

\item{denominator1}{The denominator for the first level for comparison}

\item{numerator2}{The numerator for the second level for comparison (e.g. the exposed group)}

\item{denominator2}{The denominator for the second level for comparison}
}
\value{
A dataframe with columns for rate difference, upper and lower confidence intervals
}
\description{
Calculates a 95 percent confidence interval for a rate difference.
Uses Kirkwood and Sterne's formula for a rate difference, p241 of 2nd Ed.
I think this could be re-written to work better so that it takes a data.frame as an input and compares against a given baseline.
}
\examples{
rate_difference(33, 355, 24, 518) # matches output from p241 K&S, 2nd Ed
}
