% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smokers}
\alias{smokers}
\title{Hypothetical data set from a cohort study examining the association between smoking and lung cancer}
\format{
Line list data with 90,000 rows and two variables
\describe{
 \item{smoker}{Integer; 1 for smoker, 0 for non-smoker}
 \item{lung_cancer}{Integer; 1 for lung cancer, 0 for no lung cancer}
}
}
\source{
Kirkwood, Betty R., and Jonathan AC Sterne. Essential medical statistics. John Wiley & Sons, 2010.
}
\usage{
smokers
}
\description{
Data from Kirkwood and Sterne 2nd Edition, p154
}
\keyword{datasets}
